Rename lib files to kebab-case
