+++ src/controllers/company-add.controller.js
@@ -2,7 +2,7 @@ const express = require('express')
 const Q = require('q')
 const winston = require('winston')
 const { ukOtherCompanyOptions, foreignOtherCompanyOptions } = require('../options')
-const { isBlank, toQueryString, genCSRF } = require('../lib/controllerutils')
+const { isBlank, toQueryString, genCSRF } = require('../lib/controller-utils')
 const searchService = require('../services/search.service')
 const companyService = require('../services/company.service')
 const companyFormattingService = require('../services/company-formatting.service')
