+++ src/controllers/interaction-edit.controller.js
@@ -8,7 +8,7 @@ const metadataRepository = require('../repos/metadata.repo')
 const advisorRepository = require('../repos/advisor.repo')
 const interactionDataService = require('../services/interaction-data.service')
 const interactionFormService = require('../services/interaction-form.service')
-const { genCSRF, containsFormData } = require('../lib/controllerutils')
+const { genCSRF, containsFormData } = require('../lib/controller-utils')
 const router = express.Router()
 
 function editDetails (req, res, next) {
