+++ src/controllers/interaction.controller.js
@@ -3,7 +3,7 @@ const express = require('express')
 const winston = require('winston')
 const Q = require('q')
 const interactionLabels = require('../labels/interactionlabels')
-const {genCSRF} = require('../lib/controllerutils')
+const {genCSRF} = require('../lib/controller-utils')
 const metadataRepository = require('../repos/metadata.repo')
 const interactionDataService = require('../services/interaction-data.service')
 const {getDisplayInteraction} = require('../services/interaction-formatting.service')
