+++ src/controllers/service-delivery.controller.js
@@ -3,8 +3,8 @@ const express = require('express')
 const winston = require('winston')
 const Q = require('q')
 const serviceDeliverylabels = require('../labels/servicedelivery')
-const { genCSRF, transformV2Errors } = require('../lib/controllerutils')
-const { nullEmptyFields, deleteNulls } = require('../lib/propertyhelpers')
+const { genCSRF, transformV2Errors } = require('../lib/controller-utils')
+const { nullEmptyFields, deleteNulls } = require('../lib/property-helpers')
 const metadataRepository = require('../repos/metadata.repo')
 const serviceDeliveryRepository = require('../repos/service-delivery.repo')
 const serviceDeliveryService = require('../services/service-delivery.service')
