+++ src/javascripts/clipped-list.js
@@ -1,4 +1,4 @@
-const elementStuff = require('../lib/elementstuff')
+const elementStuff = require('../lib/element-stuff')
 
 const HIDE = 'hide'
 const SHOW = 'show'
