+++ src/javascripts/company-investment-tab-form.js
@@ -1,4 +1,4 @@
-const { addClass, removeClass } = require('../lib/elementstuff')
+const { addClass, removeClass } = require('../lib/element-stuff')
 const foreignOwnedCountryWrapper = document.getElementById('foreign-owned-country-wrapper')
 const investmentAccountManagerWapper = document.getElementById('investment-account-manager-wrapper')
 const foreignOwnedRadio = document.getElementById('foreign-owned-radio')
