+++ src/lib/address.js
@@ -1,4 +1,4 @@
-const {titleCase} = require('./textformatting')
+const {titleCase} = require('./text-formatting')
 
 function getFormattedAddress (object, key) {
   if (key) {
