+++ src/lib/date-diff.js
