+++ src/lib/property-helpers.js
