+++ src/lib/text-formatting.js
