+++ src/lib/transform-sectors.js
