+++ src/repos/company.repo.js
@@ -1,7 +1,7 @@
 /* eslint camelcase: 0, prefer-promise-reject-errors: 0 */
 const Q = require('q')
 const config = require('../config')
-const authorisedRequest = require('../lib/authorisedrequest')
+const authorisedRequest = require('../lib/authorised-request')
 
 // Get a company and then pad out the interactions with related data
 function getDitCompany (token, id) {
