+++ src/repos/contact.repo.js
@@ -1,7 +1,7 @@
 const winston = require('winston')
 const Q = require('q')
 
-const authorisedRequest = require('../lib/authorisedrequest')
+const authorisedRequest = require('../lib/authorised-request')
 const config = require('../config')
 
 function getContact (token, contactId) {
