+++ src/services/contact-formatting.service.js
@@ -1,5 +1,5 @@
 const {getFormattedAddress} = require('../lib/address')
-const {newlineToBr} = require('../lib/textformatting')
+const {newlineToBr} = require('../lib/text-formatting')
 const {formatMediumDate} = require('../lib/date')
 const {formatPhone} = require('../lib/phone')
