+++ src/services/interaction-form.service.js
@@ -1,6 +1,6 @@
 const Q = require('q')
 const interactionRepository = require('../repos/interaction.repo')
-const { getPropertyId, nullEmptyFields } = require('../lib/propertyhelpers')
+const { getPropertyId, nullEmptyFields } = require('../lib/property-helpers')
 
 /**
  * Accepts an API interaction object and converts it into a format compatible with a HTML form
