+++ test/lib/controller-utils.test.js
@@ -1,6 +1,6 @@
 /* globals expect: true, describe: true, it: true */
 /* eslint no-unused-expressions: 0 */
-const { isBlank, transformV2Errors } = require('../../src/lib/controllerutils')
+const { isBlank, transformV2Errors } = require('../../src/lib/controller-utils')
 
 describe('isBlank', function () {
   it('should detects undefined variables', function () {
