+++ test/lib/property-helpers.test.js
@@ -1,7 +1,7 @@
 /* globals expect: true, describe: true, it: true */
 /* eslint no-unused-expressions: 0 */
 
-const propertyHelpers = require('../../src/lib/propertyhelpers')
+const propertyHelpers = require('../../src/lib/property-helpers')
 
 describe('PropertyHelpers: Conversion of empty values to nulls in an object with nullEmptyFields', function () {
   it('Should convert an empty string to a null', function () {
